Restructured file system by creating util package.
Moved embed util functions into util.embed
Moved member util functions into util.member
