+++ test/util/test_get_embed_with_title.py
@@ -6,7 +6,7 @@ from aiounittest.mock import AsyncMockIterator
 
 # noinspection PyUnresolvedReferences
 import test.context
-from src.util import get_embed_with_title
+from src.util.embed import get_embed_with_title
 
 with warnings.catch_warnings():
     """
